Fix metadata handling
The 'lang' metadata field for time locale selection was read only from the
page's own metadata, while accumulated metadata from meta.yaml files needs
to be taken into account as well.

Due to this change, it is now possible to set "url" fields in meta.yaml, and
it affects all relevant pages. This will lead to an error message,
if there are clashes in page urls.